Gemini With Langchain 3/requirements.txt
ipykernel
langchain
python-dotenv
langchain-google-genai
beautifulsoup4
langchain-community
faiss-cpu
streamlit
chromadb
langchain_chroma
beautifulsoup4
arxiv
pymupdf
pandas
langchain-Huggingface
